Rename dbUtils.ts to database.ts
