+++ src/utils/database.ts
@@ -1,6 +1,6 @@
 import { Pool, PoolClient } from 'pg';
 
-import { ApiError } from '../utils';  // TODO: Don't use ./utils.ts because of ./index-debug.ts
+import { ApiError } from './utils';  // TODO: Don't use ./utils.ts because of ./index-debug.ts
 import { SpraxAPIdbCfg, UserAgent, Skin, MinecraftUser, Cape, CapeType, MinecraftProfile } from '../global';
 
 export class dbUtils {
